🎨 birthday and life expectancy pickers are now reusable components
